+++ package.json
@@ -2,10 +2,10 @@
   "name": "record-actions-duration",
   "description": "",
   "version": "0.0.1",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "vitest run",
-    "bundle": "ncc build index.js --license license.txt",
+    "bundle": "ncc build src/index.js --license license.txt",
     "act": "npm run bundle && act --secret-file .secrets -j test-record-actions-duration"
   },
   "repository": {
