+++ src/durations.js
