+++ vite.config.ts
@@ -2,7 +2,10 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    include: ['test/**/*.test.js'],
-    globals: true
+    include: ['__tests__/**/*.test.js'],
+    globals: true,
+    alias: {
+      '@': '/src'
+    }
   },
 })
